VS2017 support:
NOTE: Due to a bug in Microsoft's 15.5 version of the VS2017 installer, after installing VS2017, you will need to manually add (preferably permanently) a definition for the "VS150COMNTOOLS" variable to your environment variables after installing VS2017.  The installer fails to do so for you due to a bug.

The value I had to define for my system was:

   %ProgramFiles(x86)%\Microsoft Visual Studio\2017\Community\Common7\Tools\