Don't try to do 64-bit builds on 32-bit i686 machines
Add a new build type 'i686' and don't try to do a 64-bit build
for it.  This is similar to what we do for 32-bit ARM.

Added commentary to the extpkgs.sh.ini to make it a little
easier for people editing that file to know what the choices
are.